This is a prototype of smart ventilator.
A ventilator is a machine that provides mechanical ventilation by moving breathable air into and out of the lungs, to deliver breaths to a patient who is physically unable to breathe, or breathing insufficiently. In this project we are making a prototype of smart ventilator that sense the pulse rate of the patient and automatically provide desired amount of oxygen required. It is an arduino based project.